+++ src/routes/ManageAds/states/index.ts
@@ -1,0 +1,2 @@
+export { adsDataState } from './adsDataState'
+export { adsFilterIndexState } from './adsFilterIndexState'
