+++ src/states/index.ts
@@ -1,2 +1,1 @@
-export { adsFilterIndexState } from './adsFilterIndexState'
-export { adsDataState } from './adsDataState'
+export {}
